Add detail page for search result
In past, the result of search was a list of objects of model "Courses".
This was modified, and now the search page will show objects of model
"Section". Also, clicking each row of the result table leads user to
detail page of the section in the clicked row.

Miscellaneous: add migrations folder.