Add 'default' attribute in 'Timetable' model
'default' attribute means a timetable is a timetable that users will see
at first when they click their page for timetables. Only one timetable
can be the default timetable among users' timetables. To implement this,
some methods are also added.

The followings are the added methods.

1. default()
Getter method of 'default'.

2. default(value)
Setter method of 'default'. However, this will raise AttributeError
because it can harm the meaning of 'default' attribute, which means
there must be only one default timetable.

3. set_as_default()
This method is kind of setter of 'default' attribute. It sets 'default'
attribute of the current default timetable as false, and sets the
attribute of the new default timetable as true.