Add signals for constructor and destructor of 'Timetable' instance
To make a newly added timetable as the default timetable for display,
add a signal for constructor of 'Timetable' instance. Also, to prevent
a situation that there are no default timetable when the default
timetable is deleted, add a signal for destructor of 'Timetable'
instance. To implement the destructor, 'modified_date' field is added,
which means the date when a timetable is modified.

1. set_new_timetable_as_default(sender, instance, **kwargs)
Call 'set_as_default' method of the timetable instance which is newly
added.

2. set_recent_timetable_as_default_when_default_deleted(
    sender, instance, **kwargs)
When the default timetable is deleted, the recently modified timetable
will be the default timetable of the user.

3. modified_date
A field of Timetable model which shows date and time of modification.