Add a url pattern for multiple timetables
Previously, only the first timetable of a user was displayed. Now, by
url containing the id of a timetable that the user wants to see, user
can choose the timetable they want to see. Of course, a list of
timetables connected with the web page of each timetable will be added
later.

1. get_timetable_to_display()
Return a timetable to display. If there is no matched timetable, it
returns 'None'.