Users can add and delete their timetables
Using ajax, users can now add and delete their timetables. To add a
timetable, client sends user info and the title of the new timetable.
To delete a timetable, client sends the id of the timetable. On both
cases, server will provide a proper url to redirect to client. Two
methods, 'add_new_timetable' and 'remove_timtable', in 'views.py' file
manage request for adding and deleting timetables with ajax.